move ServerAPI and Plugin into their own files, rearrange
