+++ packages/server-api/src/features.ts
@@ -1,5 +1,6 @@
-import { SignalKApiId } from '.'
-
+/**
+ * @ignore this is extended by {@link ServerAPI}, no need to document separately
+ */
 export interface Features {
   /**
    * Returns the available APIs and Plugins.
@@ -61,3 +62,13 @@ export interface FeatureInfo {
     enabled: boolean
   }>
 }
+
+export type SignalKApiId =
+  | 'resources'
+  | 'course'
+  | 'history'
+  | 'autopilot'
+  | 'anchor'
+  | 'logbook'
+  | 'historyplayback' //https://signalk.org/specification/1.7.0/doc/streaming_api.html#history-playback
+  | 'historysnapshot' //https://signalk.org/specification/1.7.0/doc/rest_api.html#history-snapshot-retrieval
